# Optimize string concat in mpd parser