@pukkandan 2022-10-08
Does this really matter? I am not interested in making millisecond optimizations