@FlxB2 2022-10-09
I think it would make a difference, initialization and media can get quite long, e.g. in this [example](https://cloudflarestream.com/31c9291ab41fac05471db4e73aa11717/manifest/video.mpd). Every time `+=` is called the interpreter has to create a new string and copy the values over, the execution time differs quite a bit. You can find more information to that [here](https://python.plainenglish.io/concatenating-strings-efficiently-in-python-9bfc8e8d6f6e).

But if you are not interested in those kind of changes I can run the tool directly on youtube-dl as well and see if the change makes sense there.