@pukkandan 2022-10-09
You didn't think to actually try timing it before "optimizing"? Here, I'll do it for you. The example manifest you gave is actually quite small, and not a suitable candidate for testing. So I took the best of 5 with the gigantic manifest in https://github.com/yt-dlp/yt-dlp/issues/3390 instead

master
```
❯ yt-dlp -s https://www.youtube.com/watch?v=mivcDoTjpQE
[youtube] mivcDoTjpQE: Downloading webpage
[youtube] mivcDoTjpQE: Downloading android player API JSON
[youtube] mivcDoTjpQE: Downloading MPD manifest
Function _parse_mpd_formats_and_subtitles Took 1.1475 seconds
[youtube] mivcDoTjpQE: Downloading MPD manifest
Function _parse_mpd_formats_and_subtitles Took 2.4663 seconds
[info] mivcDoTjpQE: Downloading 1 format(s): 18
```
PR
```
❯ yt-dlp -s https://www.youtube.com/watch?v=mivcDoTjpQE
[youtube] mivcDoTjpQE: Downloading webpage
[youtube] mivcDoTjpQE: Downloading android player API JSON
[youtube] mivcDoTjpQE: Downloading MPD manifest
Function _parse_mpd_formats_and_subtitles Took 1.1729 seconds
[youtube] mivcDoTjpQE: Downloading MPD manifest
Function _parse_mpd_formats_and_subtitles Took 2.9814 seconds
[info] mivcDoTjpQE: Downloading 1 format(s): 18
```

Your code actually ends up being marginally slower, though the difference is actually below statistical significance

Please don't waste maintainers' time with pointless PRs. If you really wanted to help improve performance, there are many places using sub-optimal code (Eg: https://github.com/yt-dlp/yt-dlp/pull/3234). But to identify and improve them, you'd have to actually put in some effort