@FlxB2 2022-10-14
Unfortunately I could not reproduce the results for your example. Seems like the inline `prepare_template` method does not get executed as far as I can tell. 

But I did execute some tests as well, unfortunately a significant difference can only be achieved for pretty long strings. 
Here is the tests I ran if your interested:

```python 
import timeit
import random
import string
import functools

# setup
d = ''.join(random.choice(string.ascii_uppercase + string.digits + string.punctuation) for _ in range(10000))

def a(tmpl):
    in_template = False
    t = ''
    for c in tmpl:
        t += c
        if c == '$':
            in_template = not in_template
        elif c == '%' and not in_template:
            t += c

def b(tmpl):
    t_list = []
    in_template = False

    for c in tmpl:
        t_list.append(c)
        if c == '$':
            in_template = not in_template
        elif c == '%' and not in_template:
            t_list.append(c)
    t = ''.join(t_list)

t_a = timeit.Timer(functools.partial(a, d))
t_b = timeit.Timer(functools.partial(b, d))
print(t_a.timeit(100))
print(t_b.timeit(100))
```
Results:
```bash
> python3 yt-dlp-test.py 
0.1256352269999752
0.11093434099996102
> pypy3 yt-dlp-test.py 
0.2907702929999232
0.06776938899997731
```

I searched for mpd files with long initialization / media attributes, but none of them were longer than 500 chars. Therefore I agree that this change does not really make sense.